requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
distlib==0.3.5
fastapi==0.80.0
filelock==3.7.1
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.30
outcome==1.2.0
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.20.1
pycparser==2.21
pydantic==1.9.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.19.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.3
virtualenv==20.16.3
webdriver-manager==3.8.3
wsproto==1.1.0
